+++ prow/BUILD.bazel
@@ -116,12 +116,10 @@ filegroup(
         "//prow/external-plugins/needs-rebase:all-srcs",
         "//prow/external-plugins/refresh:all-srcs",
         "//prow/flagutil:all-srcs",
-        "//prow/gcs/reporter:all-srcs",
         "//prow/gcsupload:all-srcs",
         "//prow/genfiles:all-srcs",
         "//prow/gerrit/adapter:all-srcs",
         "//prow/gerrit/client:all-srcs",
-        "//prow/gerrit/reporter:all-srcs",
         "//prow/git:all-srcs",
         "//prow/gitattributes:all-srcs",
         "//prow/github:all-srcs",
