+++ prow/cmd/crier/BUILD.bazel
@@ -12,17 +12,17 @@ go_library(
         "//prow/config:go_default_library",
         "//prow/config/secret:go_default_library",
         "//prow/crier:go_default_library",
+        "//prow/crier/reporters/gcs:go_default_library",
+        "//prow/crier/reporters/gerrit:go_default_library",
+        "//prow/crier/reporters/github:go_default_library",
+        "//prow/crier/reporters/slack:go_default_library",
         "//prow/flagutil:go_default_library",
-        "//prow/gcs/reporter:go_default_library",
         "//prow/gerrit/client:go_default_library",
-        "//prow/gerrit/reporter:go_default_library",
-        "//prow/github/reporter:go_default_library",
         "//prow/interrupts:go_default_library",
         "//prow/kube:go_default_library",
         "//prow/logrusutil:go_default_library",
         "//prow/pjutil:go_default_library",
         "//prow/pubsub/reporter:go_default_library",
-        "//prow/slack/reporter:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_google_cloud_go//storage:go_default_library",
         "@org_golang_google_api//option:go_default_library",
