+++ prow/crier/reporters/github/reporter_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package reporter
+package github
 
 import (
 	"testing"
