+++ prow/gerrit/adapter/BUILD.bazel
@@ -12,8 +12,8 @@ go_library(
         "//prow/apis/prowjobs/v1:go_default_library",
         "//prow/client/clientset/versioned/typed/prowjobs/v1:go_default_library",
         "//prow/config:go_default_library",
+        "//prow/crier/reporters/gerrit:go_default_library",
         "//prow/gerrit/client:go_default_library",
-        "//prow/gerrit/reporter:go_default_library",
         "//prow/pjutil:go_default_library",
         "@com_github_andygrunwald_go_gerrit//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
