+++ prow/github/BUILD.bazel
@@ -56,7 +56,6 @@ filegroup(
         ":package-srcs",
         "//prow/github/fakegithub:all-srcs",
         "//prow/github/report:all-srcs",
-        "//prow/github/reporter:all-srcs",
     ],
     tags = ["automanaged"],
 )
