+++ prow/pjutil/BUILD.bazel
@@ -22,9 +22,9 @@ go_library(
     deps = [
         "//prow/apis/prowjobs/v1:go_default_library",
         "//prow/config:go_default_library",
+        "//prow/crier/reporters/github:go_default_library",
         "//prow/gcsupload:go_default_library",
         "//prow/github:go_default_library",
-        "//prow/github/reporter:go_default_library",
         "//prow/interrupts:go_default_library",
         "//prow/kube:go_default_library",
         "//prow/pod-utils/decorate:go_default_library",
