+++ prow/plank/BUILD.bazel
@@ -13,8 +13,8 @@ go_test(
     deps = [
         "//prow/apis/prowjobs/v1:go_default_library",
         "//prow/config:go_default_library",
+        "//prow/crier/reporters/github:go_default_library",
         "//prow/github:go_default_library",
-        "//prow/github/reporter:go_default_library",
         "//prow/pjutil:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@io_k8s_api//core/v1:go_default_library",
@@ -35,9 +35,9 @@ go_library(
     deps = [
         "//prow/apis/prowjobs/v1:go_default_library",
         "//prow/config:go_default_library",
+        "//prow/crier/reporters/github:go_default_library",
         "//prow/github:go_default_library",
         "//prow/github/report:go_default_library",
-        "//prow/github/reporter:go_default_library",
         "//prow/kube:go_default_library",
         "//prow/pjutil:go_default_library",
         "//prow/pod-utils/decorate:go_default_library",
